pyproject.toml
[project]
name = "moebot"
version = "0.1.0"
description = "Default template for PDM package"
readme = "README.md"
requires-python = "==3.12.*"
authors = [
    {name = "haukuen", email = "haukuen@foxmail.com"},
]
dependencies = [
    "nonebot2>=2.3.2",
    "httpx>=0.27.0",
    "langid>=1.1.6",
    "nonebot-adapter-qq>=1.5.0",
]
license = {text = "MIT"}

[tool.nonebot]
adapters = [
    { name = "QQ", module_name = "nonebot.adapters.qq" }
]
plugins = []
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.pdm]
distribution = false

[[tool.pdm.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
verify_ssl = true
